requirements.txt
# Core Streamlit dependencies
streamlit>=1.28.0

# Snowflake connectivity
snowflake-snowpark-python>=1.9.0
snowflake-connector-python>=3.5.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.0

# Additional utilities
requests>=2.31.0